Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

RegularizedProblems.FirstOrderModel

RegularizedProblems.FirstOrderNLSModel

RegularizedProblems.MIT_matrix_completion_model

RegularizedProblems.bpdn_model

RegularizedProblems.fh_model

RegularizedProblems.group_lasso_model

RegularizedProblems.random_matrix_completion_model

​

RegularizedProblems.FirstOrderModel — Type

model = FirstOrderModel(f, ∇f!; name = "first-order model")

A simple subtype of AbstractNLPModel to represent a smooth objective.

Arguments

f :: F <: Function: a function such that f(x) returns the objective value at x;

∇f! :: G <: Function: a function such that ∇f!(g, x) stores the gradient of the objective at x in g;

x :: AbstractVector: an initial guess.

Keyword arguments

selected :: AbstractVector{<: Int}: a list of variables to apply the regularizer to (default: all variables).

All other keyword arguments are passed through to the NLPModelMeta constructor.

RegularizedProblems.FirstOrderNLSModel — Type

model = FirstOrderNLSModel(r!, jv!, jtv!; name = "first-order NLS model")

A simple subtype of AbstractNLSModel to represent a nonlinear least-squares problem with a smooth residual.

Arguments

r! :: R <: Function: a function such that r!(y, x) stores the residual at x in y;

jv! :: J <: Function: a function such that jv!(u, x, v) stores the product between the residual Jacobian at x and the vector v in u;

jtv! :: Jt <: Function: a function such that jtv!(u, x, v) stores the product between the transpose of the residual Jacobian at x and the vector v in u;

x :: AbstractVector: an initial guess.

Keyword arguments

selected :: AbstractVector{<: Int}: a list of variables to apply the regularizer to (default: all variables).

All other keyword arguments are passed through to the NLPModelMeta constructor.

RegularizedProblems.MIT_matrix_completion_model — Method

model, nls_model, sol = MIT_matrix_completion_model(args...)

A special case of matrix completion problem in which the exact image is a noisy MIT logo.

See the documentation of random_matrix_completion_model() for more information.

RegularizedProblems.bpdn_model — Method

model, nls_model, sol = bpdn_model(args...)
model, nls_model, sol = bpdn_model(compound = 1, args...)

Return an instance of an NLPModel and an instance of an NLSModel representing the same basis-pursuit denoise problem, i.e., the under-determined linear least-squares objective

½ ‖Ax - b‖₂²,

where A has orthonormal rows and b = A * x̄ + ϵ, x̄ is sparse and ϵ is a noise vector following a normal distribution with mean zero and standard deviation σ.

Arguments

m :: Int: the number of rows of A

n :: Int: the number of columns of A (with n ≥ m)

k :: Int: the number of nonzero elements in x̄

noise :: Float64: noise standard deviation σ (default: 0.01).

The second form calls the first form with arguments

m = 200 * compound
n = 512 * compound
k =  10 * compound

Keyword arguments

bounds :: Bool: whether or not to include nonnegativity bounds in the model (default: false).

Return Value

An instance of a FirstOrderModel and of a FirstOrderNLSModel that represent the same basis-pursuit denoise problem, and the exact solution x̄.

If bounds == true, the positive part of x̄ is returned.

RegularizedProblems.fh_model — Method

fh_model(; kwargs...)

Return an instance of an NLPModel and an instance of an NLSModel representing the same Fitzhugh-Nagumo problem, i.e., the over-determined nonlinear least-squares objective

½ ‖F(x)‖₂²,

where F: ℝ⁵ → ℝ²⁰² represents the fitting error between a simulation of the Fitzhugh-Nagumo model with parameters x and a simulation of the Van der Pol oscillator with fixed, but unknown, parameters.

Keyword Arguments

All keyword arguments are passed directly to the ADNLPModel (or ADNLSModel) constructure, e.g., to set the automatic differentiation backend.

Return Value

An instance of an ADNLPModel that represents the Fitzhugh-Nagumo problem, an instance of an ADNLSModel that represents the same problem, and the exact solution.

RegularizedProblems.group_lasso_model — Method

model, nls_model, sol = group_lasso_model(args...)
model, nls_model, sol = group_lasso_model(compound = 1, args...)

Return an instance of an NLPModel and NLSModel representing the group-lasso problem, i.e., the under-determined linear least-squares objective

½ ‖Ax - b‖₂²,

where A has orthonormal rows and b = A * x̄ + ϵ, x̄ is sparse and ϵ is a noise vector following a normal distribution with mean zero and standard deviation σ. Note that with this format, all groups have a the same number of elements and the number of groups divides evenly into the total number of elements.

Arguments

m :: Int: the number of rows of A

n :: Int: the number of columns of A (with n ≥ m)

g :: Int : the number of groups

ag :: Array{Int}: group-index denoting which groups are active (with max(ag) ≤ g), i.e. [1, 4, 5] when there are 7 groups

noise :: Float64: noise amount ϵ (default: 0.01).

The second form calls the first form with arguments

m = 200 * compound
n = 512 * compound
k =  10 * compound

Return Value

An instance of a FirstOrderModel that represents the group-lasso problem. An instance of a FirstOrderNLSModel that represents the group-lasso problem. Also returns true x, number of groups g, group-index denoting which groups are active, and a Matrix where rows are group indices of x

RegularizedProblems.random_matrix_completion_model — Method

model, nls_model, sol = random_matrix_completion_model(args...)

Return an instance of an NLPModel and an instance of an NLSModel representing the same matrix completion problem, i.e., the square linear least-squares objective

½ ‖P(X - A)‖²

in the Frobenius norm, where X is the unknown image represented as an m x n matrix, A is a fixed image, and the operator P only retains a certain subset of pixels of X and A.

Arguments

m :: Int: the number of rows of X and A

n :: Int: the number of columns of X and A

r :: Int: the desired rank of A

sr :: AbstractFloat: a threshold between 0 and 1 used to determine the set of pixels retained by the operator P

va :: AbstractFloat: the variance of a first Gaussian perturbation to be applied to A

vb :: AbstractFloat: the variance of a second Gaussian perturbation to be applied to A

c :: AbstractFloat: the coefficient of the convex combination of the two Gaussian perturbations.

Return Value

An instance of a FirstOrderModel and of a FirstOrderNLSModel that represent the same matrix completion problem, and the exact solution.